pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cairn"
version = "1.0.0"
description = "Migrate map data between OnX Backcountry and CalTopo"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Scott"}
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
cairn = "cairn.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["cairn"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
